feat(select): fraction support in --select + deterministic scaling sweep
select_top now accepts a fraction (0<n<1 = top fraction of the pool, e.g. 0.25 = top
quarter) in addition to a count (>=1) or 0 (all). --select in build_map_reduce.py is now
float-typed. Adds campaign/select_sweep.py: re-runs the REDUCE selection in memory across
select levels and reports surfaced-principle count + statement bytes/tokens — the
deterministic principle-layer scaling, no LLM, no package clobber. (Full adapter-byte
comparison still needs p2b_finish per level; the sweep notes this.)

+++ tools/subagent_factory/map_reduce_build.py
@@ -168,7 +168,7 @@ def assemble(
     embedder: Embedder,
     cos: float = 0.55,
     decisions: dict[int, dict] | None = None,
-    select: int = 0,
+    select: float = 0,
 ) -> dict:
     """Write the REDUCE'd distilled layer into subagents/<slug>/. Returns a counts summary."""
     cache_root = repo / "cache" / "book-extracts"
